:sparkles: Improve messaging group list.
