+++ lib/domain/repositories/messenger_repository.dart
@@ -2,9 +2,9 @@ import 'package:aewallet/application/wallet/wallet.dart';
 import 'package:aewallet/domain/models/core/failures.dart';
 import 'package:aewallet/domain/models/core/result.dart';
 import 'package:aewallet/model/available_networks.dart';
+import 'package:aewallet/model/data/access_recipient.dart';
 import 'package:aewallet/model/data/account.dart';
-import 'package:aewallet/model/messenger/talk.dart';
-import 'package:aewallet/model/public_key.dart';
+import 'package:aewallet/model/data/messenger/talk.dart';
 
 abstract class MessengerRepositoryInterface {
   Future<Result<List<String>, Failure>> getTalkIds();
@@ -12,8 +12,8 @@ abstract class MessengerRepositoryInterface {
   Future<Result<Talk, Failure>> getTalk(String talkId);
 
   Future<Result<Talk, Failure>> createTalk({
-    required List<PublicKey> members,
-    required List<PublicKey> admins,
+    required List<AccessRecipient> members,
+    required List<AccessRecipient> admins,
     required Account creator,
     required LoggedInSession session,
     required NetworksSetting networkSettings,
