+++ lib/ui/views/main/messenger_tab/components/public_key_item.dart
@@ -1,8 +1,8 @@
 import 'package:aewallet/application/settings/settings.dart';
 import 'package:aewallet/application/settings/theme.dart';
+import 'package:aewallet/model/data/access_recipient.dart';
 import 'package:aewallet/ui/util/access_recipient_formatters.dart';
 import 'package:aewallet/ui/util/styles.dart';
-import 'package:aewallet/ui/views/main/messenger_tab/bloc/create_talk.dart';
 import 'package:aewallet/ui/widgets/components/dialog.dart';
 import 'package:aewallet/ui/widgets/components/item_remove_button.dart';
 import 'package:aewallet/util/get_it_instance.dart';
