+++ lib/ui/views/main/messenger_tab/components/talk_list_item.dart
@@ -1,4 +1,8 @@
+import 'package:aewallet/application/settings/theme.dart';
+import 'package:aewallet/model/data/messenger/talk.dart';
+import 'package:aewallet/ui/util/styles.dart';
 import 'package:aewallet/ui/views/main/messenger_tab/bloc/providers.dart';
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -13,7 +17,7 @@ abstract class TalkListItem extends ConsumerWidget {
 
   const factory TalkListItem.loaded({
     Key? key,
-    required String talkId,
+    required Talk talk,
   }) = _LoadedTalkListItem;
 
   const factory TalkListItem.loading({
@@ -36,12 +40,7 @@ class _AutoloadTalkListItem extends TalkListItem {
     return asyncTalk.map(
       error: (_) => const TalkListItem.loading(),
       loading: (_) => const TalkListItem.loading(),
-      data: (talk) => Card(
-        child: SizedBox(
-          height: 48,
-          child: Text(talk.value.name),
-        ),
-      ),
+      data: (talk) => TalkListItem.loaded(talk: talk.value),
     );
   }
 }
@@ -55,30 +54,75 @@ class _LoadingTalkListItem extends TalkListItem {
   final Duration? animationDelay;
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) => const Card(
-        child: SizedBox(height: 48),
-      )
-          .animate(
-            delay: animationDelay ?? Duration.zero,
-            onPlay: (controller) => controller.repeat(),
-          )
-          .shimmer(delay: 1000.ms, angle: 0.5);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final theme = ref.watch(ThemeProviders.selectedTheme);
+
+    return Card(
+      shape: RoundedRectangleBorder(
+        side: BorderSide(
+          color: theme.backgroundAccountsListCardSelected!,
+        ),
+        borderRadius: BorderRadius.circular(10),
+      ),
+      elevation: 0,
+      color: theme.backgroundAccountsListCardSelected,
+      child: const SizedBox(height: 48),
+    )
+        .animate(
+          delay: animationDelay ?? Duration.zero,
+          onPlay: (controller) => controller.repeat(),
+        )
+        .shimmer(
+          blendMode: BlendMode.srcOver,
+          delay: 1000.ms,
+          angle: 0.5,
+        );
+  }
 }
 
 class _LoadedTalkListItem extends TalkListItem {
   const _LoadedTalkListItem({
     super.key,
-    required this.talkId,
+    required this.talk,
   });
 
-  final String talkId;
+  final Talk talk;
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) => Card(
-        key: Key(talkId),
-        child: SizedBox(
-          height: 48,
-          child: Text(talkId),
+  Widget build(BuildContext context, WidgetRef ref) {
+    final theme = ref.watch(ThemeProviders.selectedTheme);
+
+    return Card(
+      shape: RoundedRectangleBorder(
+        side: BorderSide(
+          color: theme.backgroundAccountsListCardSelected!,
         ),
-      );
+        borderRadius: BorderRadius.circular(10),
+      ),
+      elevation: 0,
+      color: theme.backgroundAccountsListCardSelected,
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 10),
+        color: theme.backgroundAccountsListCard,
+        child: Column(
+          children: [
+            Align(
+              alignment: Alignment.topLeft,
+              child: AutoSizeText(
+                talk.name,
+                style: theme.textStyleSize12W600Primary,
+              ),
+            ),
+            Align(
+              alignment: Alignment.centerRight,
+              child: AutoSizeText(
+                '${talk.members.length} members',
+                style: theme.textStyleSize10W100Primary,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
 }
